pyproject.toml
[tool.poetry]
name = "shopstyle_api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.8"
django-rest-framework = "^0.1.0"
django-extensions = "^3.1.3"
pymongo = {extras = ["srv"], version = "^3.12.0"}
django-environ = "^0.7.0"
json-lines = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
